competitors-websites/requirements.txt
Flask==1.1.1
Psycopg2-binary